add conditions for stopping

+++ deploy/down.sh
@@ -4,5 +4,16 @@ if [[ -f ./env.sh ]]; then
     . ./env.sh
 fi
 
-docker stop $(docker ps -q --filter name=$CONTAINER_NAME)
-docker rm $(docker ps -a -q --filter name=$CONTAINER_NAME)
+# Stop Running Containers
+RUNNING_CONTAINERS=$(docker ps -q --filter name=$CONTAINER_NAME)
+
+if [ ${#RUNNING_CONTAINERS} -ne 0 ]; then
+    docker stop $(docker ps -q --filter name=$CONTAINER_NAME)
+fi
+
+# Delete Containers
+ALL_CONTAINERS=$(docker ps -a -q --filter name=$CONTAINER_NAME)
+
+if [ ${#ALL_CONTAINERS} -ne 0 ]; then
+    docker rm $(docker ps -a -q --filter name=$CONTAINER_NAME)
+fi
